Drop class component.
